feat(project): :sparkles: expose typography components and types on lib
